Refactor project structure - separate helpers module and analysis
