+++ pymt/__init__.py
@@ -3,3 +3,4 @@ import toml
 
 logger = default_logger.logger
 config = toml.load("config.toml")
+selected_lines = config['pymt']['selected_lines']
