+++ pymt/pymt.py
@@ -1,8 +1,10 @@
-from pymt import default_logger, logger, influxdb_functions, mongo_functions, api, map
-from pymt import config
+from pymt import logger, api, map, default_logger
+from pymt.helpers import infl, mon
 import pymt.models.oasth as model
 import time
 import datetime
+
+influx = infl.init_influxdb()
 
 
 @default_logger.timer("Getting bus telematics...")
@@ -12,37 +14,18 @@ def get_buses(lines):
     return [model.Bus(b) for route in route_telematics for b in route if b]
 
 
-# ---------------------------------------------------------------------------------
-
-influx = influxdb_functions.init_influxdb()
-selected_lines = config['pymt']['selected_lines']
-
-
-def load_lines(days, lines=selected_lines):
-    _lines = []
-    if not lines:
-        _lines = mongo_functions.get_all_lines()
-    else:
-        _lines = [mongo_functions.get_line_by_name(line_name) for line_name in lines]
-    [logger.debug(f"Selected lines: {l.__dict__}") for l in _lines]
-    for d in days:
-        for line in _lines:
-            line.stops[d] = mongo_functions.get_route_stops(route_id=line.days[d])
-    return _lines
-
-
 def run():
     today = datetime.datetime.today().weekday()
     days = [today * 2, today * 2 + 1]
     logger.debug(days)
-    loaded_lines = load_lines(days)
+    loaded_lines = mon.load_lines(days)
 
     while True:
 
         new_day = datetime.datetime.today().weekday()
         if new_day != today:
             days = [new_day * 2, new_day * 2 + 1]
-            loaded_lines = load_lines(days)
+            loaded_lines = mon.load_lines(days)
 
         for l in loaded_lines:
             logger.debug("Line: {}".format(l.name))
@@ -52,7 +35,7 @@ def run():
         buses = get_buses(loaded_lines)
         [logger.debug(b.__dict__) for b in buses]
 
-        influxdb_functions.save_buses(influx, buses)
+        infl.save_buses(influx, buses)
 
         m = map.init_map()
         [map.plot_route(m, line, days) for line in loaded_lines]
